Checkpoint Clicker contour smoothing, raised artwork and smooth border prototype
Preserve shared smooth artwork contours and raised relief. Add the selectable Smooth border +1.5 mm prototype while retaining exact silhouette mode and compatible switch mechanics.

Clean redundant border samples and triangulate at Float32 precision to prevent collapsed STL triangles. Include regression fixtures and focused Cat/squirrel 35 mm mechanics, topology, shared-boundary and STL coverage verified under Three.js 0.167.1 and 0.185.1.

+++ clicker/image-geometry.js
@@ -71,12 +71,16 @@ export function createAccentRegionGeometries(accentRegions, autoFit, scaleMultip
 
       const mmOuter = outer.map((p) => {
         const { mmX, mmY } = pxPointToMM(p, autoFit, scaleMultiplier);
-        return { x: mmX, y: mmY };
+        // Triangulate at the precision used by BufferGeometry and binary STL.
+        // Otherwise almost-collinear doubles can form cap triangles that
+        // collapse when their vertices are stored as Float32. This produces
+        // the same final boundary coordinates, including shared color edges.
+        return { x: Math.fround(mmX), y: Math.fround(mmY) };
       });
       const mmHoles = holes.map((hole) =>
         hole.map((p) => {
           const { mmX, mmY } = pxPointToMM(p, autoFit, scaleMultiplier);
-          return { x: mmX, y: mmY };
+          return { x: Math.fround(mmX), y: Math.fround(mmY) };
         })
       );
